refactoring(ci): do not use uint #46

+++ app/apps/project/adapter/project_adapter.go
@@ -10,12 +10,12 @@ import (
 
 type ProjectInput struct {
 	Title    string `json:"title"`
-	StatusId int    `json:"status_id"`
+	StatusId int64  `json:"status_id"`
 }
 
 // ProjectDto - project dto.
 type ProjectDto struct {
-	ID        uint      `json:"id"`
+	ID        int64     `json:"id"`
 	Title     string    `json:"title"`
 	Code      string    `json:"code"`
 	CreatedAt time.Time `json:"created_at"`
