+++ app/internal/token/auth_token_service.go
@@ -10,13 +10,13 @@ import (
 
 type ApiAuthService interface {
 	// Login generates new access + refresh tokens and stores refresh token
-	Login(ctx context.Context, userID uint, device string) (*AccessTokenResult, *RefreshTokenResult, error)
+	Login(ctx context.Context, userID int64, device string) (*AccessTokenResult, *RefreshTokenResult, error)
 
 	// Logout deletes refresh token (single device)
-	Logout(ctx context.Context, userID uint, refreshToken string) error
+	Logout(ctx context.Context, userID int64, refreshToken string) error
 
 	// ListSessions returns all refresh tokens for a user
-	ListSessions(ctx context.Context, userID uint) ([]*db.UserToken, error)
+	ListSessions(ctx context.Context, userID int64) ([]*db.UserToken, error)
 
 	// RefreshAccessToken validates refresh token and returns a new access token
 	RefreshAccessToken(ctx context.Context, refreshToken string) (*AccessTokenResult, error)
@@ -34,7 +34,7 @@ type AuthService struct {
 	refreshTokenService RefreshTokenService
 }
 
-func (s *AuthService) Login(ctx context.Context, userID uint, device string) (*AccessTokenResult, *RefreshTokenResult, error) {
+func (s *AuthService) Login(ctx context.Context, userID int64, device string) (*AccessTokenResult, *RefreshTokenResult, error) {
 	accessResult, err := s.accessTokenService.Create(userID, AudTokenAPIUser)
 	if err != nil {
 		return nil, nil, err
@@ -58,12 +58,12 @@ func (s *AuthService) Login(ctx context.Context, userID uint, device string) (*A
 	return accessResult, refreshResults, nil
 }
 
-func (s *AuthService) Logout(ctx context.Context, userID uint, refreshToken string) error {
+func (s *AuthService) Logout(ctx context.Context, userID int64, refreshToken string) error {
 	connection := db.GetDb(ctx)
 	return db.DeleteTokenByRefreshString(ctx, connection, userID, refreshToken)
 }
 
-func (s *AuthService) ListSessions(ctx context.Context, userID uint) ([]*db.UserToken, error) {
+func (s *AuthService) ListSessions(ctx context.Context, userID int64) ([]*db.UserToken, error) {
 	tokens, err := db.GetTokens(ctx, db.GetDb(ctx), userID)
 	if err != nil {
 		return nil, err
